Feat: 상품 카테고리로 조회     - 주요데이터 모두 `page.tsx`로 이동
